# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations": {
      "post": {
        "operationId": "OrganizationsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create organization",
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Organizations retrieved."
          }
        },
        "summary": "Get all organizations",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization retrieved."
          }
        },
        "summary": "Get organization by ID",
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update organization",
        "tags": [
          "organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/search-user": {
      "post": {
        "operationId": "OrganizationsController_searchUser_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search user by email",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/members": {
      "post": {
        "operationId": "OrganizationsController_addMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member added."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add member to organization",
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_getMembers_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get organization members",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/members/{userId}": {
      "delete": {
        "operationId": "OrganizationsController_removeMember_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove member from organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/members/leave": {
      "post": {
        "operationId": "OrganizationsController_leaveMembership_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Left organization."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/users": {
      "get": {
        "operationId": "OrganizationsController_getOrganizationUsers_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get organization users by role",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/users/{userId}/role": {
      "patch": {
        "operationId": "OrganizationsController_updateMemberRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member role updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update member role",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/users/{userId}": {
      "get": {
        "operationId": "OrganizationsController_getMemberDetails_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member details retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get member details",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/roles": {
      "get": {
        "operationId": "OrganizationsController_getRoles_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roles retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get organization roles",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_createRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create organization role",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/roles/{roleId}": {
      "patch": {
        "operationId": "OrganizationsController_updateRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update organization role",
        "tags": [
          "organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_deleteRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete organization role",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/invitations/accept": {
      "post": {
        "operationId": "OrganizationsController_acceptInvitation_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation accepted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept organization invitation",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/courses": {
      "get": {
        "operationId": "OrganizationsController_getOrganizationCourses_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "termId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "levelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instructor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPublished",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses retrieved."
          }
        },
        "summary": "Get organization courses",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "OrganizationsController_createOrganizationCourse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create organization course",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/courses/{courseId}": {
      "patch": {
        "operationId": "OrganizationsController_updateOrganizationCourse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update organization course",
        "tags": [
          "organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_deleteOrganizationCourse_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete organization course",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/organizations/{id}/courses/{courseId}/assign-term": {
      "patch": {
        "operationId": "OrganizationsController_assignCourseToTerm_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTermDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course assigned to term."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign course to term",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/v1/levels/organization/{organizationId}": {
      "get": {
        "operationId": "LevelController_findByOrganization_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Level"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shadboard API",
    "description": "The Shadboard API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          }
        }
      },
      "SearchUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "levelId": {
            "type": "string"
          },
          "termId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "email",
          "roleId"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          },
          "levelId": {
            "type": "string"
          },
          "termId": {
            "type": "string"
          }
        },
        "required": [
          "roleId"
        ]
      },
      "RolePermissions": {
        "type": "object",
        "properties": {}
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/RolePermissions"
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/RolePermissions"
          }
        }
      },
      "CreateOrganizationCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isOrgPrivate": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "description",
          "level",
          "category",
          "price"
        ]
      },
      "UpdateOrganizationCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean"
          },
          "isOrgPrivate": {
            "type": "boolean"
          }
        }
      },
      "AssignTermDto": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string"
          }
        },
        "required": [
          "termId"
        ]
      },
...
    }
  }
}